New code for sensivities and adaptive integration
NEW: vde.py
First parts of an implementation of the variational differential equations
for the calculation of accurate and efficient sensitivities.
Sensivities are needed to provide a jacobian
of the function we would like to optimize.

NEW: integrators.py
First functions for possible more adapted integration
with adaptive or even moving grid.