setup class changed
Constructor now takes a full InitialDistribution object
instead of just the initial distribution function.
This is because we need more and more derivatives of
the initial distribution when implementing VDEs or adaptive integration approaches.
In addition, it takes jacobians of the drift and diffusion functions
w.r.t. the Fokker-Planck parameters.

Fixed a false calculation of the grid's step sizes.

Setup.Integration has additional parameters now for adaptive integration.
They will be for sure changed later, plans for adaptive integration are not yet as clear.

The generators make_rhsFcn, make_jacFcn now take
a new parameters "sensivities" to indicate whether the ODE should be
augmented by the variational differential equations.
RHS is implemented, jacobian not yet.